Move following code from user model to following concern
